+++ paper/paper.md
@@ -20,9 +20,9 @@ authors:
     orcid: 0000-0002-7856-1399
     affiliation: 2
 affiliations:
- - name: Department of Plant and Environmental Sciences, Clemson University
+ - name: Pee Dee Research and Education Center, Department of Plant and Environmental Sciences, Clemson University, USA
    index: 1
- - name: Biological and Environmental Science and Engineering Division, King Abdullah University of Science and Technology
+ - name: Biological and Environmental Science and Engineering Division and Center for Desert Agriculture, King Abdullah University of Science and Technology, Thuwal, Saudi Arabia
    index: 2
 date: 7 January 2025
 bibliography: paper.bib
