+++ .github/workflows/update-docs.yml
@@ -16,4 +16,4 @@ jobs:
         uses: peaceiris/actions-gh-pages@v3
         with:
           publish_dir: ./docs 
-          github_token: ${{ secrets.ACTIONS_PAT }}
+          github_token: ${{ secrets.GITHUB_TOKEN }}
