### 2025-05-04 #7632 @majensen
Hi @sportiellomike -  would you be willing to review [this work](#7632] for [JOSS](https://joss.theoj.org)? Based your past JOSS contribution, your expertise would be welcome. Thanks for considering --@majensen
