### 2025-05-23 #8263 @editorialbot
[REVIEW]: Coordinate: An Android application to organize samples into grids
**Submitting author:** @trife (<a href="http://orcid.org/0000-0002-5974-6523">Trevor Rife</a>)
**Repository:** https://github.com/PhenoApps/Coordinate
**Branch with paper.md** (empty if default branch):
**Version:** v4.1
