### 2025-05-30 #8263 @sportiellomike
Thanks for working hard on the software and for making it open source @trife. I've begun my review, and I have been creating issues on the [repo](https://github.com/PhenoApps/Coordinate/issues). I have tagged these beginning with the phrase:
> [JOSS REVIEW]

to make it obvious.

Feel free to begin working on these issues now, but I will notify you when I am finished reviewing. I have not yet begun to evaluate certain elements, so don't take a lack of checkbox as an issue yet. @majensen let me know if there are any problems with this plan.

Thanks!
