+++ docs/export.md
@@ -1,11 +1,6 @@
-Exporting data
-==============
+# Exporting data
 
-Once you are finished collecting data, grids can be exported by
-selecting the 'Export' (<img ref="save" style="vertical-align: middle;" src="_static/icons/save.png" width="20px">)
-option from either the grid list item or, if they are a part of a
-project, the project list item.
+Once you are finished collecting data, grids can be exported by selecting the 'Export' (<img class="icon" src="_static/icons/save.png">) option from either the grid list item or, if they are a part of a project, the project list item.
 
-A default filename containing the grid name and current date is
-prefilled and can be edited prior to export. Exported grids are saved in
-CSV format and are located in the "exports" directory.
+A default filename containing the grid name and current date is prefilled and can be edited prior to export. 
+Exported grids are saved in CSV format and are located in the "exports" directory.
