+++ docs/funding.md
@@ -1,18 +1,0 @@
-Funding
-=======
-
-?> This project is under active development.
-
-Current
--------
-
-USDA NIFA No. 2019-67021-29931
-
-US Agency for International Development (USAID) No. 7200AA‐19LE‐00005
-
-Previous
---------
-
-McKnight Foundation
-
-National Science Foundation No. 1543958
