+++ docs/settings.md
@@ -1,56 +1,43 @@
-Settings
-========
+# Settings
 
-Collection
-----------
+## Collection
 
-#### Direction (<img ref="direction" style="vertical-align: middle;" src="_static/icons/direction.png" width="20px">)
+#### Direction (<img class="icon" src="_static/icons/direction.png">)
 
-By default, Coordinate moves down each column when <a href="#/collect">Collecting Data</a>. This setting allows users to move across rows (left to right).
+By default, Coordinate moves down each column when [Collecting Data](collect.md).
+This setting allows users to move across rows (left to right).
 
-#### Notification sounds (<img ref="sounds" style="vertical-align: middle;" src="_static/icons/sounds.png" width="20px">)
+#### Notification sounds (<img class="icon" src="_static/icons/sounds.png">)
 
-Enabling this setting adds a sound both when finishing a column and once
-the last cell in the grid has been filled.
+Enabling this setting adds a sound both when finishing a column and once the last cell in the grid has been filled.
 
-#### Unique values (<img ref="unique" style="vertical-align: middle;" src="_static/icons/unique.png" width="20px"> and <img ref="unique-options" style="vertical-align: middle;" src="/_static/icons/unique-options.png" width="20px">)
+#### Unique values (<img class="icon" src="_static/icons/unique.png"> and <img class="icon" src="/_static/icons/unique-options.png">)
 
-This setting allows the user to disallow duplicated values on a per grid
-basis, per project basis, or entirely within the app.
+This setting allows the user to disallow duplicated values on a per grid basis, per project basis, or entirely within the app.
 
-Layout
-------
+## Layout
 
-#### Scaling (<img ref="resize" style="vertical-align: middle;" src="/_static/icons/resize.png" width="20px">)
+#### Scaling (<img class="icon" src="/_static/icons/resize.png">)
 
-The user can use this setting adjust how large the grid UI is in the
-Collect Data screen.
+The user can use this setting adjust how large the grid UI is in the Collect Data screen.
 
-Export
-------
+## Export
 
-#### Project export (<img ref="export" style="vertical-align: middle;" src="/_static/icons/export.png" width="20px">)
+#### Project export (<img class="icon" src="/_static/icons/export.png">)
 
-<a href="#/projects">Projects</a> can be exported with each grid
-being a single file (all contained within in a zipped file) or all grids
-appended to a single CSV file.
+[Projects](projects.md) can be exported with each grid being a single file (all contained within in a zipped file) or all grids appended to a single CSV file.
 
-#### Share exports (<img ref="share" style="vertical-align: middle;" src="/_static/icons/share.png" width="20px">)
+#### Share exports (<img class="icon" src="/_static/icons/share.png">)
 
-This setting will utilize the Android share system, allowing the
-exported file to be sent using different apps already installed on the
-device.
+This setting will utilize the Android share system, allowing the exported file to be sent using different apps already installed on the device.
 
-Storage
--------
+## Storage
 
-#### Define storage location (<img ref="storage-location" style="vertical-align: middle;" src="/_static/icons/storage-location.png" width="20px">)
+#### Define storage location (<img class="icon" src="/_static/icons/storage-location.png">)
 
-This setting allows for the modification of the default Storage
-Location.
+This setting allows for the modification of the default Storage Location.
 
-#### Database (<img ref="database" style="vertical-align: middle;" src="/_static/icons/database.png" width="20px">)
+#### Database (<img class="icon" src="/_static/icons/database.png">)
 
-The internal database can be exported, imported, or deleted. Deleting
-the database will entirely erase all data that was collected and any
-other customizations that have been made within Coordinate.
+The internal database can be exported, imported, or deleted.
+Deleting the database will entirely erase all data that was collected and any other customizations that have been made within Coordinate.
