+++ docs/collect.md
@@ -1,36 +1,19 @@
-Collecting data
-===============
+# Collecting data
 
-Data Input
-----------
-
-<figure align="center" class="image">
-  <img src="_static/images/collect_data_framed.png" width="400px"> 
-  <figcaption><i>Data collection
-screen</i></figcaption> 
+<figure class="image">
+  <img class="screenshot" src="_static/images/collect_data_framed.png" width="350px"> 
+  <figcaption class="screenshot-caption"><i>Data collection screen</i></figcaption> 
 </figure>
 
-Each cell within a grid can hold a single string of data (such as a
-plant barcode). Data can be input via the text box below the grid.
-Sample barcodes can be rapidly scanned to increase data throughput using
-the barcode button next to data textbox
-(<img ref="barcode" style="vertical-align: middle;" src="_static/icons/barcode.png" width="20px">).
+To start collecting data, create and open a new [grid](grids.md).
+Each cell within a grid can hold a single string of data (such as a plant barcode).
+Data can be input via the text box below the grid.
+Sample barcodes can be rapidly scanned to increase data throughput using the barcode button next to data textbox (<img class="icon" src="_static/icons/barcode.png">).
 
-The current cell is highlighted in orange. Cells containing saved data
-turn dark blue. Cells that have been excluded from holding data in a
-given grid are grey.
+The current cell is highlighted in orange.
+Cells containing saved data turn dark blue.
+Cells that have been excluded from holding data in a given grid are grey.
 
-Metadata
---------
-
-Metadata specific to the active grid can be viewed by selecting the
-Metadata icon (<img ref="metadata" style="vertical-align: middle;" src="_static/icons/metadata.png" width="20px">)
-on the top toolbar.
-
-Project
--------
-
-The active grid can be added to an existing or new project via the
-Project Editor icon
-(<img ref="project-add" style="vertical-align: middle;" src="_static/icons/project-add.png" width="20px">) on the
-top toolbar.
+The top toolbar has additional relevant actions.
+Metadata for the active grid can be viewed by selecting the Metadata icon (<img class="icon" src="_static/icons/metadata.png">) on the top toolbar.
+The active grid can be added to an existing or new project via the Project Editor icon (<img class="icon" src="_static/icons/project-add.png">) on the top toolbar.
