+++ projects.md
@@ -1,44 +1,29 @@
-Projects
-========
+# Projects
 
-<figure align="center" class="image">
-  <img src="_static/images/project_list_framed.png" width="400px"> 
-  <figcaption><i>Project list
-layout</i></figcaption> 
+<figure class="image">
+  <img class="screenshot" src="_static/images/project_list_framed.png" width="350px"> 
+  <figcaption class="screenshot-caption"><i>Project list layout</i></figcaption> 
 </figure>
 
-Groups of grids can be organized into a single project, simplifying data
-collection for larger experiments. Projects can be exported as a single
-file or multiple files, adjustable in the <a href="#/settings">Settings</a>. Grids can be added to Projects when they\'re created or
-from the <a href="#/collect">Collecting Data</a> screen.
+Groups of grids can be organized into a single project, simplifying data collection for larger experiments. Projects can be exported as a single file or multiple files, adjustable in the [Settings](settings.md).
+Grids can be added to Projects when they\'re created or from the [Collecting Data](collect.md) screen.
 
-List Item Layout
-----------------
+## List Item Layout
 
-<figure align="center" class="image">
-  <img src="_static/images/project_list_item.png" width="600px"> 
-  <figcaption><i>Individual project list
-item</i></figcaption> 
+<figure class="image">
+  <img class="screenshot" src="_static/images/project_list_item.png" width="600px"> 
+  <figcaption class="screenshot-caption"><i>Individual project list item</i></figcaption> 
 </figure>
 
-Each project in the list contains the project name, number of grids, and
-date created on the left.
+Each project in the list contains the project name, number of grids, and date created on the left.
 
-Action buttons on the right of each list item allow a project to be
-deleted (<img ref="delete" style="vertical-align: middle;" src="_static/icons/delete.png" width="20px">), exported
-(<img ref="save" style="vertical-align: middle;" src="_static/icons/save.png" width="20px">), grids in the project
-to be viewed (<img ref="grid-list" style="vertical-align: middle;" src="_static/icons/grid-list.png" width="20px">),
-or new grids to be created from that project
-(<img ref="grid-add" style="vertical-align: middle;" src="_static/icons/grid-add.png" width="20px">).
+Action buttons on the right of each list item allow a project to be deleted (<img class="icon" src="_static/icons/delete.png">), exported (<img class="icon" src="_static/icons/save.png">), grids in the project to be viewed (<img class="icon" src="_static/icons/grid-list.png">), or new grids to be created from that project (<img class="icon" src="_static/icons/grid-add.png">).
 
-New Projects
-------------
+## New Projects
 
-New projects can be created by clicking the New Project icon
-(<img ref="project-new" style="vertical-align: middle;" src="_static/icons/project-new.png" width="20px">) on the
-top toolbar. New projects only require a name to be created.
+New projects can be created by clicking the New Project icon (<img class="icon" src="_static/icons/project-new.png">) on the top toolbar.
+New projects only require a name to be created.
 
-Deleting Projects
------------------
+## Deleting Projects
 
 Deleting a project will delete all of the grids within that project.
