+++ templates.md
@@ -1,49 +1,32 @@
-Templates
-=========
+# Templates
 
-<figure align="center" class="image">
-  <img src="_static/images/template_list_framed.png" width="400px"> 
-  <figcaption><i>Template list layout</i></figcaption> 
+<figure class="image">
+  <img class="screenshot" src="_static/images/template_list_framed.png" width="400px"> 
+  <figcaption class="screenshot-caption"><i>Template list layout</i></figcaption> 
 </figure>
 
-Templates allow defined structure and metadata across different grids
-that might be collected. Two templates are included by default, a 6x20
-Seed Tray Template and a 8x12 DNA Plate template.
+Templates allow defined structure and metadata across different grids that might be collected. Two templates are included by default, a 6x20 Seed Tray Template and a 8x12 DNA Plate template.
 
-List Item Layout
-----------------
+## List Item Layout
 
-<figure align="center" class="image">
-  <img src="_static/images/template_list_item.png" width="600px"> 
-  <figcaption><i>Individual template list item</i></figcaption> 
+<figure class="image">
+  <img class="screenshot" src="_static/images/template_list_item.png" width="600px"> 
+  <figcaption class="screenshot-caption"><i>Individual template list item</i></figcaption> 
 </figure>
 
-Each template in the list contains the template name, size, and date
-created on the left.
+Each template in the list contains the template name, size, and date created on the left.
 
-Action buttons on the right of each template list item allow non-default
-templates to be deleted
-(<img ref="delete" style="vertical-align: middle;" src="/_static/icons/delete.png" width="20px">), edited
-(<img ref="template-edit" style="vertical-align: middle;" src="/_static/icons/template-edit.png" width="20px">), exported
-(<img ref="save" style="vertical-align: middle;" src="/_static/icons/save.png" width="20px">), grids created from
-that template to be viewed
-(<img ref="grid-list" style="vertical-align: middle;" src="/_static/icons/grid-list.png" width="20px">), or new grids
-to be created from that template
-(<img ref="grid-add" style="vertical-align: middle;" src="/_static/icons/grid-add.png" width="20px">).
+Action buttons on the right of each template list item allow non-default templates to be deleted (<img class="icon" src="/_static/icons/delete.png">), edited (<img class="icon" src="/_static/icons/template-edit.png">), exported (<img class="icon" src="/_static/icons/save.png">), grids created from that template to be viewed (<img class="icon" src="/_static/icons/grid-list.png">), or new grids to be created from that template (<img class="icon" src="/_static/icons/grid-add.png">).
 
-New Templates
--------------
+## New Templates
 
-Templates can be created by selecting the New Template icon
-(<img ref="template-add" style="vertical-align: middle;" src="/_static/icons/template-add.png" width="20px">) or
-imported by selecting the Import icon
-(<img ref="import" style="vertical-align: middle;" src="/_static/icons/import.png" width="20px">) on the top toolbar.
-Creating a new template opens a step-wise process that allows for
-additional customization.
+Templates can be created by selecting the New Template icon (<img class="icon" src="/_static/icons/template-add.png">) or imported by selecting the Import icon
+(<img class="icon" src="/_static/icons/import.png">) on the top toolbar.
+Creating a new template opens a step-wise process that allows for additional customization.
 
-<figure align="center" class="image">
-  <img src="_static/images/template_new_process.png" width="1000px"> 
-  <figcaption><i>Template creation process</i></figcaption> 
+<figure class="image">
+  <img class="screenshot" src="_static/images/template_new_process.png" width="1000px"> 
+  <figcaption class="screenshot-caption"><i>Template creation process</i></figcaption> 
 </figure>
 
 1.  The name and dimensions of the template are defined
@@ -55,14 +38,11 @@ additional customization.
[…]

[…]
