### 2025-06-19 edd0c90
add example applications
