### 2025-07-20 #8263 @sportiellomike
@majensen I hope to finish my review soon. I'm working on acquiring an Android phone to actually test the functionality, as I do not own one myself. It should be done soon.
