### 2025-07-28 #8263 @majensen
@svburke -- would you have interest in reviewing [this paper](#8263) for JOSS, or know someone who might? As @trife [notes](https://github.com/openjournals/joss-reviews/issues/8263#issuecomment-3097175213), you can set up an Android emulator to test. Let me know -
