### 2025-07-29 #8263 @sportiellomike
@majensen I have finished my review. I recommend accept given the quality of the work submitted, the completion of the checklist, and the quick communication of the authors. Thank you for the opportunity to engage with JOSS and the open source community.

@trife Thanks for contributing to the open source community. I did use an emulator back in June, but wanted to verify the scanning function of your app worked well. I apologize it took me so long to actually get my hands on an Android phone with a camera. I left some open issues on your github that are just nitpicky preferences, and I hope do not interfere with your acceptance to JOSS.
