### 2025-08-09 #8263 @svburke
@majensen : I have finished my review and checklist. I recommend `accept` given the quality of the repository code, the well written manuscript, and the usefulness of the application. I have one issue that I created, but it is far from being blocker, more of a minor inconvenience. Thank you for the opportunity to review for JOSS.

@trife : Thank you for your efforts and valuable contributions to the open-source community. I found the Coordinate application intuitive and easy to use from setup through input and export. Its customization makes it adaptable to a wide range of scientific disciplines involving sample management, and I appreciate the simplicity and effectiveness of its interface.
