### 2025-08-12 c2940f8
fixed edit text delete not working moving to an entry will put cursor at end of text now
