### 2025-08-12 25f62f3
Add optional tutorial during setup
