### 2025-08-22 #8263 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
