### 2025-08-22 7a285e7
Fix formatting and punctuation in paper.md
