+++ paper/paper.md
@@ -30,19 +30,19 @@ bibliography: paper.bib
 
 # Summary
 
-Efficient and accurate sample organization is one of the most important aspects of modern biological science. `Coordinate` is an Android application that is used to organize sample identifiers into virtual grids. Sample identifiers are collected via text or barcode into individual cells that are structured in a Cartesian coordinate grid. Grids inherit properties from user-defined templates including their dimensions, row and column naming, metadata fields, and any specific cells to exclude from data entry. Grids are exported as CSV files that are easily integrated into downstream analysis pipelines since metadata fields are user-defined. Grids can also be grouped into projects and exported as a single ZIP file. `Coordinate` can be further customized with options to change the direction of cell procession for data input, add notification sounds, input random blank cells for quality control, and adjust the level of uniqueness required for incoming sample identifiers. Two default templates are included with `Coordinate`: DNA Plate and Seed Tray.
+Efficient and accurate sample organization is one of the most important aspects of modern biological science. `Coordinate` is an Android application that is used to organize sample identifiers into virtual grids. Sample identifiers are collected via text or barcode into individual cells that are structured in a Cartesian coordinate grid. Grids inherit properties from user-defined templates including their dimensions, row and column naming, metadata fields, and any specific cells to exclude from data entry. Grids are exported as CSV files that are easily integrated into downstream analysis pipelines since metadata fields are user-defined. Grids can also be grouped into projects and exported as a single ZIP file. `Coordinate` can be further customized with options to change the direction of cell procession for data input, add notification sounds, input random blank cells for quality control, and adjust the level of uniqueness required for incoming sample identifiers. Two default templates are included with `Coordinate`: `DNA Plate` and `Seed Tray`.
 
-Although other tools exist to design grids for activities like tissue collection [@iastateTissueSample], `Coordinate` takes a unique approach of "sampling as you go" which does not assign specific samples to specific cells before the samples are collected. This approach can help to minimize mistakes since matching an identifier is not required before collecting a sample and missing samples do not need to be manually replaced in the sampling interface. Further, the customizability of `Coordinate` allows it to be adapted to many different use cases or needs.
+Although other tools exist to design grids for activities like tissue collection [@iastateTissueSample], `Coordinate` takes a unique approach of "sampling as you go", which does not assign specific samples to specific cells before the samples are collected. This approach can help to minimize mistakes since matching an identifier is not required before collecting a sample and missing samples do not need to be manually replaced in the sampling interface. Furthermore, the customizability of `Coordinate` allows it to be adapted to many different use cases or needs.
 
 # Statement of need
 
-Statistical power increases with the number of samples being used for analysis, but managing samples at scale requires optimized tools and worfklows to minimize mistakes. Organizational errors impede scientific progress and can lead to erroneous results or publication retractions [@Ying:2013;@Lee:2013;@Moloney:2017]. Many life science research activities organize samples into grids and require metadata to be collected for downstream analyses. `Coordinate` enables robust and customizable grid-based data collection by creating a mobile, virtual grid layout that can be populated as samples are collected. In contrast to unstructured approaches, `Coordinate` provides standardized data workflows that reduce mistakes and allow for more efficient sample collection.
[…]
