### 2025-09-01 c54b363
Add compose, init screens, use stateflow in projects
